requirements.txt
aiocfscrape==1.0.0
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.15.0
certifi==2024.6.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
Js2Py==0.74
loguru==0.7.2
magic-filter==1.0.12
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.3
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.7.3
pydantic-settings==2.3.1
pydantic_core==2.18.4
Pygments==2.18.0
pyjsparser==2.7.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-socks==2.5.0
PyYAML==6.0.1
redis==5.0.5
rich==13.7.1
rq==1.16.2
rsa==4.9
schedule==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.30
starlette==0.37.2
Telethon==1.36.0
TgCrypto==1.2.5
typer==0.12.3
typing_extensions==4.12.1
tzlocal==5.2
ujson==5.10.0
unipayment==1.1.2
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
